refactor(driver): ♻️ extract pipeline stages into driver/pipeline.h
