+++ compiler/driver/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(daoc main.cpp)
+add_executable(daoc main.cpp pipeline.cpp)
 target_link_libraries(daoc PRIVATE dao_analysis dao_backend_llvm dao_ir dao_frontend)
 
 # Bake the path to the runtime static library so `daoc build` can link it.
